LayoutTests/http/tests/security/cross-frame-access-object-getPrototypeOf-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-object-getPrototypeOf-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-object-getPrototypeOf.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-object-getPrototypeOf-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-object-getPrototypeOf.html. Domains, protocols and ports must match.

This tests that you can't get the prototype of the window or history objects cross-origin using Object.getPrototypeOf().

PASS: Object.getPrototypeOf(targetWindow) should be 'undefined' and is.
PASS: Object.getPrototypeOf(targetWindow.history) should be 'undefined' and is.
